move checkpoint back to mpc-node and simplify
