+++ chain-signatures/node/src/protocol/request/task.rs
@@ -76,7 +76,7 @@ impl GeneratingPhase {
         } else {
             PendingPresignature::InStorage(
                 self.presignature_id,
-                mpc_primitives::Participant(u32::from(self.proposer)),
+                self.proposer,
                 ctx.presignatures.clone(),
             )
         };
@@ -85,13 +85,10 @@ impl GeneratingPhase {
         let gen_ctx = ctx.generate_ctx();
         let generator = match SignGenerator::new(
             &gen_ctx,
-            mpc_primitives::Participant(u32::from(self.proposer)),
+            self.proposer,
             Arc::clone(&state.request),
             presignature_pending,
-            self.accepted_participants
-                .iter()
-                .map(|p| mpc_primitives::Participant(u32::from(*p)))
-                .collect(),
+            self.accepted_participants.clone(),
         )
         .await
         {
