+++ chain-signatures/node/src/sign_bidirectional.rs
@@ -1,14 +1,59 @@
 use crate::protocol::{Chain, IndexedSignRequest};
 use alloy::primitives::{keccak256, Address, Bytes};
+use cait_sith::protocol::Participant;
 use k256::elliptic_curve::point::AffineCoordinates;
 use k256::elliptic_curve::sec1::ToEncodedPoint as _;
 use k256::{AffinePoint, Scalar};
 use mpc_crypto::derive_key;
-pub use mpc_primitives::{
-    BidirectionalTx, ChainFromError, Participant, PublishState, SignBidirectionalEvent, SignStatus,
-    Signature,
-};
+pub use mpc_primitives::{BidirectionalTx, ChainFromError, SignBidirectionalEvent, Signature};
 use rlp::{Rlp, RlpStream};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct PublishState {
+    pub signature: Signature,
+    pub participants: Vec<Participant>,
+    pub is_proposer: bool,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub enum SignStatus {
+    PendingGeneration,
+    PendingPublish { publish: PublishState },
+    PendingExecution { tx: BidirectionalTx },
+    PendingGenerationBidirectional,
+    PendingPublishBidirectional { publish: PublishState },
+}
+
+impl SignStatus {
+    pub fn is_pending_generation(&self) -> bool {
+        matches!(
+            self,
+            SignStatus::PendingGeneration | SignStatus::PendingGenerationBidirectional
+        )
+    }
+
+    pub fn is_pending_execution(&self) -> bool {
+        matches!(self, SignStatus::PendingExecution { .. })
+    }
+
+    /// Project this status onto what is observable at a checkpoint's own chain height.
+    pub fn consensus_tag(&self) -> u8 {
+        match self {
+            SignStatus::PendingGeneration | SignStatus::PendingPublish { .. } => 0,
+            SignStatus::PendingExecution { .. } => 1,
+            SignStatus::PendingGenerationBidirectional
+            | SignStatus::PendingPublishBidirectional { .. } => 2,
+        }
+    }
+
+    pub fn execution_tx(&self) -> Option<&BidirectionalTx> {
+        match self {
+            SignStatus::PendingExecution { tx } => Some(tx),
+            _ => None,
+        }
+    }
+}
 
 pub type RequestId = [u8; 32];
