+++ integration-tests/src/lib.rs
@@ -25,11 +25,12 @@ use mpc_chain_ethereum::EthConfig;
 use mpc_chain_solana::SolConfig;
 use mpc_contract::config::{PresignatureConfig, ProtocolConfig, TripleConfig};
 use mpc_contract::primitives::CandidateInfo;
+use mpc_node::backlog::Checkpoint;
 use mpc_node::gcp::GcpService;
 use mpc_node::indexer_hydration::HydrationConfig;
 use mpc_node::web::CheckpointResponse;
 use mpc_node::{logs, mesh, node_client, storage};
-use mpc_primitives::{Chain, Checkpoint};
+use mpc_primitives::Chain;
 use near_workspaces::network::Sandbox;
 use near_workspaces::types::{KeyType, SecretKey};
 use near_workspaces::{Account, AccountId, Contract, Worker};
